~ due to the way the EVE protocol is designed only one instance of the same account can be logged in at the same time, so in order for everyone to properly communicate changes have been made to ConnectionManager in the Cluster to properly work that way ~ the LoginQueue no longer sets up the session data, that's the responsability of the ClientConnection ~ updated ID usages to longs for account and other things ~ due to the last change, some packets have been altered in the way they are encoded/decoded, the Marshal library really could benefit from a rewrite to handle integers and other objects in a better way + added opening connection support to EVEClientSocket for it to be used in the Node ~ rewritten Node connection to tidy up things a bit, the code should be easier to follow now ~ first steps into getting rid of static objects in Program.cs with the new NodeContainer

+++ Server/ClusterController/NodeConnection.cs
@@ -8,9 +8,9 @@ namespace ClusterControler
 {
     public class NodeConnection: Connection
     {
-        public int NodeID { get; private set; }
+        public long NodeID { get; private set; }
         
-        public NodeConnection(EVEClientSocket socket, ConnectionManager connectionManager, int nodeID)
+        public NodeConnection(EVEClientSocket socket, ConnectionManager connectionManager, long nodeID)
             : base(socket, connectionManager)
         {
             this.NodeID = nodeID;
